Revert "Clean archi"
