--- lib/browse/presentation/pages/browse_view.dart
@@ -1,10 +1,0 @@
-import 'package:flutter/material.dart';
-
-class BrowseView extends StatelessWidget {
-  const BrowseView({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return const Placeholder();
-  }
-}
