+++ lib/core/configurations/app_router.dart
@@ -1,9 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:movies_app/core/configurations/pages_routes.dart';
 
-import '../../home/presentation/pages/movie_details_view.dart';
-import '../../splash/splash_view.dart';
-import '../layout_view.dart';
+import '../../views/main_layout/layout_view.dart';
+import '../../views/splash/splash_view.dart';
 
 class AppRouter {
   static Route<dynamic> onGenerateRoute(RouteSettings settings) {
@@ -15,10 +14,6 @@ class AppRouter {
       case PagesRoutes.layoutView:
         return MaterialPageRoute(
             builder: (context) => LayoutView(), settings: settings);
-
-      case PagesRoutes.movieDetailsView:
-        return MaterialPageRoute(
-            builder: (context) => MovieDetailsView(), settings: settings);
 
       default:
         return MaterialPageRoute(
