+++ lib/core/configurations/pages_routes.dart
@@ -2,5 +2,4 @@
 class PagesRoutes {
   static const String initial = "/";
   static const String layoutView = "layout_view";
-  static const String movieDetailsView = "movie_details_view";
 }
