+++ lib/main.dart
@@ -2,7 +2,6 @@ import 'package:flutter/material.dart';
 
 import 'core/configurations/app_router.dart';
 import 'core/configurations/pages_routes.dart';
-import 'core/theme/application_theme_manager.dart';
 
 void main() {
   runApp(const MyApp());
@@ -14,9 +13,9 @@ class MyApp extends StatelessWidget {
   // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
+    return const MaterialApp(
         debugShowCheckedModeBanner: false,
-        theme: ApplicationThemeManager.theme,
+        //theme: ApplicationThemeManager.darkModeTheme
         initialRoute: PagesRoutes.initial,
         onGenerateRoute: AppRouter.onGenerateRoute);
   }
