+++ lib/views/search/search_view.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class SearchView extends StatelessWidget {
+  const SearchView({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Placeholder();
+  }
+}
