+++ lib/views/watchlist/watchlist_view.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class WatchlistView extends StatelessWidget {
+  const WatchlistView({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Placeholder();
+  }
+}
